+++ packages/dart_flutter_common/lib/list.dart
@@ -1,0 +1,8 @@
+extension SafeIndexAccess<T> on List<T> {
+  T? safeGet(int index) {
+    if (index >= 0 && index < length) {
+      return this[index];
+    }
+    return null;
+  }
+}
